requirements.txt
ultralytics>=8.0.0
matplotlib>=3.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
seaborn>=0.11.0
pyyaml>=6.0
torch>=2.0.0
torchvision>=0.15.0
onnx>=1.12.0
onnxruntime>=1.8.0
protobuf>=3.20.0 